GUI Update #2
The basics of the GUI have been completed at this point.  We just need
to figure out what the best way to present the data to the users is, and
then implement the containers and methods to get that done.  The next
step will probably be working with HtmlUnit to get information from the
website.